Fix circular imports Add unit tests of `locations` module Refactor modules for clarity
