+++ ovos_config/config.py
@@ -18,7 +18,7 @@ from time import sleep
 
 from mycroft_bus_client import Message
 
-from ovos_config.base import LocalConf, MycroftDefaultConfig, MycroftSystemConfig, MycroftUserConfig, RemoteConf
+from ovos_config.models import LocalConf, MycroftDefaultConfig, MycroftSystemConfig, MycroftUserConfig, RemoteConf
 from ovos_config.locations import OLD_USER_CONFIG, get_xdg_config_save_path, get_xdg_config_locations
 from ovos_config.utils import FileWatcher
